+++ src/main/java/com/example/employeeapp1/Repository/EmployeePayrollRepository.java
@@ -5,5 +5,5 @@ package com.example.employeeapp1.Repository;
 import com.example.employeeapp1.Model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeePayrollRepository extends JpaRepository<Employee, Long> {
+public interface EmployeePayrollRepository extends JpaRepository<Employee, Integer> {
 }
